requirements.txt
pyautogui
requests
Pillow
pygame
google-genai